use arch packages if possible

+++ zshrc
@@ -7,18 +7,30 @@ zle -N self-insert url-quote-magic
 #############################
 # Load plugins
 #############################
-# enable fuzzy finder if it exists
-if ! [[ -f ~/.fzf.zsh ]] ; then
-  if ! [[ -f ~/.fzf/install ]] ; then
-    rm -rf ~/.fzf
-    git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
+if [[ -f /etc/arch-release ]]; then
+  package_list=(fzf diff-so-fancy zsh-syntax-highlighting zsh-completions)
+  if ! pacman -Qq ${package_list[@]} > /dev/null 2>&1; then
+    echo "You're using Archlinux! Use pacman to manage some nice tools:"
+    echo ${package_list[@]}
   fi
-  ~/.fzf/install --all
 fi
 
-source ~/.fzf.zsh
-export FZF_DEFAULT_OPTS="-m --cycle"
+if [[ -x /usr/bin/fzf && -f /usr/share/fzf/key-bindings.zsh &&
+    -f /usr/share/fzf/completion.zsh ]]; then
+  source /usr/share/fzf/key-bindings.zsh
+  source /usr/share/fzf/completion.zsh
+else
+  if ! [[ -f ~/.fzf.zsh ]] ; then
+    if ! [[ -f ~/.fzf/install ]] ; then
+      rm -rf ~/.fzf
+      git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
+    fi
+    ~/.fzf/install --all
+  fi
+  source ~/.fzf.zsh
+fi
 
+export FZF_DEFAULT_OPTS="-m --cycle"
 if (( $+commands[rg] )) ; then
   export FZF_DEFAULT_COMMAND='rg -l ""'
 elif (( $+commands[ag] )); then
@@ -32,20 +44,26 @@ zstyle :zplug:tag depth 1
 source "${HOME}/.zplug/init.zsh"
 
 zplug "zplug/zplug"
-zplug "zsh-users/zsh-completions", use:"*.plugin.zsh"
 zplug "mafredri/zsh-async", use:"async.zsh"
 zplug "leomao/zsh-hooks", use:"*.plugin.zsh"
 zplug "leomao/vim.zsh", use:vim.zsh, defer:1
 zplug "leomao/pika-prompt", use:pika-prompt.zsh, defer:2
 
-zplug "so-fancy/diff-so-fancy", as:command, use:diff-so-fancy
-
 export ENHANCD_DISABLE_HOME=1
 export ENHANCD_DOT_ARG='.'
 zplug "b4b4r07/enhancd", use:init.sh
 
+if [[ -f /etc/arch-release ]]; then
+  if [[ -d /usr/share/zsh/plugins/zsh-syntax-highlighting ]]; then
+    zplug "/usr/share/zsh/plugins/zsh-syntax-highlighting", from:local, defer:3
+  fi
+else
+  zplug "zsh-users/zsh-completions", use:"*.plugin.zsh"
+  zplug "so-fancy/diff-so-fancy", as:command, use:diff-so-fancy
+  zplug "zsh-users/zsh-syntax-highlighting", use:"*.plugin.zsh", defer:3
+fi
+
 export ZSH_HIGHLIGHT_HIGHLIGHTERS=(main brackets)
-zplug "zsh-users/zsh-syntax-highlighting", use:"*.plugin.zsh", defer:3
 
 if ! zplug check --verbose; then
   zplug install
